requirements.txt
torch>=2.0.0
pyro-ppl>=1.9.0
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0
scikit-learn>=1.3.0
pyyaml>=6.0
tqdm>=4.65.0
requests>=2.28.0
falconn>=1.3.0
